amazing refactors with david
