+++ scripts/runUnusedExports.ts
@@ -1,0 +1,7 @@
+import {
+    findUnusedExports,
+    logUnusedExportsToConsole
+} from "./findUnusedExports.js"
+
+const unusedExports = findUnusedExports()
+logUnusedExportsToConsole(unusedExports)
